IAMExtDevice::Calibrate updated
git-svn-id: https://svn.code.sf.net/p/directshownet/code/trunk@783 b2a09c84-0b36-484d-b36e-ffab392acda3

+++ directshowlib/Test/v1.3/IAMExtDeviceTest.cs
@@ -115,7 +115,7 @@ namespace DirectShowLib.Test
 			hr = pEvent.GetEventHandle(out eventHandle);
 			DsError.ThrowExceptionForHR(hr);
 
-			hr = _extDevice.Calibrate(eventHandle, (int)ExtDeviceCaps.Active, out ret);
+			hr = _extDevice.Calibrate(eventHandle, ExtDeviceCaps.Active, out ret);
 			//DsError.ThrowExceptionForHR(hr); //E_NOTIMPL , but atleast it's called.
 
 			hr = _extDevice.Calibrate(eventHandle, 0, out ret);
